+++ Sekoia.io/CHANGELOG.md
@@ -13,6 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add an action to add events to a case
 
+## 2024-12-11 - 2.65.11
+
+### Changed
+
+- Update synchronize asset action
+
+## 2024-11-27 - 2.65.7
+
+### Changed
+
+- Update get alert action with an option to get the associated cases
+
 ## 2024-11-27 - 2.65.4
 
 ### Changed